Refactored to create more generic helper functions and move logic to transformation method
